refactor: change predictions structure, fix typo file, and date
