--- src/prediction/prediction.module.ts
@@ -1,9 +1,0 @@
-import { Module } from "@nestjs/common";
-import { PredictionController } from "./prediction.controller";
-import { PredictionService } from "./prediction.service";
-@Module({
-    controllers: [PredictionController],
-    providers: [PredictionService],
-})
-
-export class PredictionModule {}
