+++ src/users/users.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
-import { ModelModule } from 'src/module/model.module';
+import { ModelModule } from 'src/model/model.module';
 import { FirestoreModule } from 'src/firestore/firestore.module';
 
 @Module({
